// src/core/enums/users-routing.enum.ts
export enum UsersRoutingEnum {
  REGISTRATION = 'registration',
  LOGIN = 'login',
  PERSONAL_INFORMATION = 'personal-information',
  SEND_EMAIL_APPROVE = 'send-email-approve',
  APPROVE_EMAIL = 'approve-email',
  GET_WAITING_USERS_FOR_APPROVE = 'get-waiting-users-for-approve',
  APPROVE_USERS_IDENTITY = 'approve-users-identity',
}